Publicly expose fmt_fuchsia_backtrace even in no_std
